pull download button into its own component
