+++ src/frontend/src/views/Data/components/SamplesView/components/SampleTableModalManager/components/DownloadModal/components/DownloadButton/style.ts
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+import { Button, CommonThemeProps, getSpaces } from "czifui";
+
+export const StyledButton = styled(Button)`
+  max-width: fit-content;
+
+  ${(props: CommonThemeProps) => {
+    const spaces = getSpaces(props);
+    return `
+      margin-top: ${spaces?.xxl}px;
+    `;
+  }}
+`;
